Fix bug causing 'ml_get: cannot find line {N}'
After deleting lines and re-entering text at the same lines, this error
was occasionally seen and would hang Vim. The issue was due to multiple
GhostNotify calls occurring (the buffer being updated in the middle of
other events).

The fix is to have GhostNotify block until the buffer is updated. The
main thread triggers an event to the socket thread to read the updates
from Vim (this part did not change). Once this happens, the socket
thread then triggers an event back to the main thread indicating the
read is complete (this part is new). GhostNotify waits on this event
before proceeding.