Fix bug when trying to send data to Broken pipe
This issue can be re-created by launching GhostText and quickly closing
it on a large file (data is still being received and the close frame
message is missing). Any updates to Vim will cause the plugin to try to
send data through what is now a broken pipe.

This commit fixes the issue by marking the thread as invalid and causing
it to finish.