Fix hang on second launch of GhostText
The second time GhostText is launched, the first socket is still marked
as valid and deadlocks since it is completed and no longer polling for
events.

This commit fixes the bug by marking the websocket as invalid when it
finishes serving and then has all invalid sockets removed from the
HTTPSERVER list every GhostNotify update.