Fix bug when receiving large frames from browser
Occasionally when receiving a large frame from the browser, an invalid
opcode or incorrect number of payload bytes are received. Trying to
re-read the socket returns errno 107 (Transport endpoint is not
connected).

This commit fixes the bug by marking the frame as invalid and continuing
the polling process so that a subsequent valid frame is received (on the
next data sent from the browser).